flask_app/src/requirements.txt
Flask==3.1.2
flask-sqlalchemy==3.1.1 
SQLAlchemy-Utils==0.42.1
PyMySQL==1.1.2
flask-login==0.6.3
Flask-Admin==1.6.1
WTForms==3.1.2
Flask-WTF==1.2.2
flask-wtf==1.2.2
python-dotenv==1.2.1
flake8==7.3.0

pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
coverage==7.3.2